fix: autoLinkPlugin 를 TypeScript로 변환
