+++ src/components/memo-editor/LexicalMarkdownEditor/LexicalMarkdownEditor.tsx
@@ -24,7 +24,7 @@ import type { EditorState } from 'lexical';
 import { useCallback } from 'react';
 import { lexicalTheme } from './LexicalMarkdownEditor.theme';
 import './LexicalMarkdownEditor.toolbar.css';
-import AutoLinkPlugin from './plugins/AutoLinkPlugin.js';
+import AutoLinkPlugin from './plugins/AutoLinkPlugin';
 import CodeHighlightPlugin from './plugins/CodeHighlightPlugin';
 import ListMaxIndentLevelPlugin from './plugins/ListMaxIndentLevelPlugin';
 import ToolbarPlugin from './plugins/ToolbarPlugin';
@@ -88,6 +88,20 @@ export default function LexicalMarkdownEditor({
     [onChange],
   );
 
+  const handleContentEditableClick = (e: React.MouseEvent<HTMLDivElement>) => {
+    const target = e.target as HTMLElement;
+    const targetParent = target.closest('a') || target.parentElement?.closest('a');
+    console.log('target', e);
+    if (targetParent?.tagName === 'A') {
+      const href = (targetParent as HTMLAnchorElement).href;
+      if (href) {
+        window.open(href, '_blank');
+        e.preventDefault();
+        e.stopPropagation();
+      }
+    }
+  };
+
   return (
     <div className="sm:h-[80%] h-[87%] w-full">
       <LexicalComposer initialConfig={initialConfig}>
@@ -99,6 +113,7 @@ export default function LexicalMarkdownEditor({
                 className={`flex-1 p-4 outline-none resize-none border border-gray-300 overflow-y-auto focus:border-blue-500 focus:ring-1 focus:ring-blue-500 ${
                   isMobile ? 'pb-20' : ''
                 }`}
+                onClick={handleContentEditableClick}
               />
             }
             placeholder={
